Added PnPHttpProvider for Microsoft Graph SDK.
